Switch from long to int to store memory, required for IBM
git-svn-id: https://svn-wrf-var.cgd.ucar.edu/trunk@738 70815b0f-b814-0410-81f7-dfa10a885da6

+++ da_3dvar/src/da_tracing/da_memory.c
@@ -15,7 +15,7 @@ void da_memory_(
 
   memory_used)
 
-  long *memory_used;
+  int *memory_used;
 {
 #ifndef MACOS
 struct mallinfo result;
